requirements.txt
boto3
pyyaml=5.3.1
svn==1.0.1
